Conflicts:
	io.h

+++ fallback.cpp
@@ -915,7 +915,7 @@ int wcwidth( wchar_t c )
 #endif
 
 #ifndef HAVE_WCSNDUP
-wchar_t *wcsndup( const wchar_t *in, int c )
+wchar_t *wcsndup( const wchar_t *in, size_t c )
 {
 	wchar_t *res = (wchar_t *)malloc( sizeof(wchar_t)*(c+1) );
 	if( res == 0 )
@@ -1111,7 +1111,7 @@ int lrand48_r(struct drand48_data *buffer, long int *result)
 
 int srand48_r(long int seedval, struct drand48_data *buffer)
 {
-	buffer->seed = (int)seedval;
+	buffer->seed = (unsigned int)seedval;
 	return 0;
 }
